ラッパーの共通部をコモンクラス化してやったZE
