+++ src/main/java/jp/sasrai/biomepainter/util/wrapper/BiomeBase_1_9.java
@@ -9,51 +9,15 @@ import java.lang.reflect.Method;
 /**
  * Created by sasrai on 2017/03/04.
  */
-public class BiomeBase_1_9 extends WrapperBase implements BiomeBaseInterface {
-    private final Class<?> biomeBase;
-    private Object instance;
-
-    public BiomeBase_1_9() {
-        biomeBase = getNMSBiomeBase();
-    }
-
-    private Class<?> getNMSBiomeBase() {
-        try {
-            String biomebasePackage = getNMSPackage() + "BiomeBase";
-            return Class.forName(biomebasePackage);
-        } catch (Exception e) {
-            return null;
-        }
-    }
-
+public class BiomeBase_1_9 extends BiomeBaseCommonWrapper {
     private Method getIDMethod() throws NoSuchMethodException {
-        return biomeBase.getMethod("a", biomeBase);
+        return getBiomeBaseClass().getMethod("a", getBiomeBaseClass());
     }
     private int getIDMethodData() {
         try {
-            instance.toString();
-            return (int) getIDMethod().invoke(null, instance);
+            return (int) getIDMethod().invoke(null, getBiomeBaseObject());
         } catch (NoSuchMethodException | IllegalAccessException | InvocationTargetException e) {
             return -1;
-        }
-    }
-
-    private boolean isBBNameField(Field field) {
-        return (field.getType() == String.class);
-    }
-    private Field getNameField() throws NoSuchFieldException {
-        for (Field field: biomeBase.getDeclaredFields()) {
-            if (isBBNameField(field)) {
-                return field;
-            }
-        }
-        throw new NoSuchFieldException();
-    }
-    private String getNameFieldData() {
-        try {
-            return (String)getNameField().get(instance);
-        } catch (NoSuchFieldException | IllegalAccessException e) {
-            return "OCEAN";
         }
     }
 
@@ -79,30 +43,5 @@ public class BiomeBase_1_9 extends WrapperBase implements BiomeBaseInterface {
     @Override
     public int getId() {
         return getIDMethodData();
-    }
-
-    @Override
-    public String getName() {
-        return getNameFieldData();
-    }
-
-    @Override
-    public String getCanonicalName() {
-        return biomeBase.getCanonicalName();
-    }
-
-    @Override
-    public Class<?> getBiomeBaseClass() {
-        return biomeBase;
-    }
-
-    @Override
-    public Object getBiomeBaseObject() {
-        return instance;
-    }
-
-    @Override
-    public void setBiomeBaseObject(Object biomeBaseObject) {
-        instance = biomeBaseObject;
     }
 }
